+++ data-transfer-object/src/main/java/org/grnet/cat/dtos/assessment/AssessmentResponse.java
@@ -76,4 +76,13 @@ public class AssessmentResponse {
     )
     @JsonProperty("parent_assessment_id")
     public String parentAssessmentId;
+
+    @Schema(
+            type = SchemaType.STRING,
+            implementation = String.class,
+            description = "The parent of the assessment",
+            example = "c242e43f-9869-4fb0-b881-631bc5746ec0"
+    )
+    @JsonProperty("assessment_doc_version")
+    public String version;
 }